requirements.txt
Django==4.1.5
djangorestframework==3.14.0
drf-yasg==1.21.4
Pillow==9.4.0
django-celery-results==2.4.0
celery==5.2.7
redis==4.4.2
psycopg2==2.9.5
psycopg2-binary==2.9.5